+++ index.html
@@ -230,6 +230,7 @@
     <div class="img-upload__message  img-upload__message--loading">Загружаем...</div>
   </template>
 
-  <script src="js/main.js"></script>
+  <script src="js/script.js" type="module"></script>
+
 </body>
 </html>
