+++ js/script.js
@@ -1,0 +1,3 @@
+import './util.js';
+import './comments.js';
+import './photo-description.js';
